Refactoring

+++ src/common/LTimer.hpp
@@ -1,21 +1,3 @@
-/* Copyright (C) 2015  Fredrik Mörtberg <fredrikmo@hotmail.com>
- *
- * This file is a part of the Hunta2 project.
- *
- * Hunta2 is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Hunta2 is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Hunta2.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 #ifndef LTIMER_H
 #define LTIMER_H
 
@@ -53,7 +35,7 @@ public:
     /**
      * Will return the time that the timer has been running.
      */
-    Uint32 getTicks();
+    Uint64 getTicks();
 
     /**
      * Check if the timer is started.
@@ -68,8 +50,8 @@ public:
     virtual ~LTimer();
 protected:
 private:
-    Uint32 mStartTicks;
-    Uint32 mPausedTicks;
+    Uint64 mStartTicks;
+    Uint64 mPausedTicks;
 
     bool mPaused;
     bool mStarted;
